+++ docs/changelog.rst
@@ -23,6 +23,7 @@ Changes in 0.9.X - DEV
 - Document save raise an exception if save_condition fails #1005
 - Fixes some internal _id handling issue. #961
 - Updated URL and Email Field regex validators, added schemes argument to URLField validation. #652
+- Removed get_or_create() deprecated since 0.8.0. #300
 
 Changes in 0.9.0
 ================